Vite migration

See merge request tadeas0/midi-search!41

+++ backend/src/app/search_engine/strategy/standardization_strategy.py
@@ -16,7 +16,7 @@ class DefaultStrategy(StandardizationStrategy):
 
 class RelativeIntervalStrategy(StandardizationStrategy):
     def standardize(self, top_line: npt.NDArray[np.int64]) -> npt.NDArray[np.int64]:
-        res = np.zeros(len(top_line))
+        res = np.zeros(len(top_line), dtype=np.int64)
         for i in range(1, len(top_line)):
             res[i] = top_line[i] - top_line[i - 1]
         return res
